fix  (core): Make EnolaApplicationTest.docGenTemplatedGreetingN() pass
It's *CORRECT* like this (and has been for a while) - because if we
only --load greetingN.ttl, then rdfs: etc. links should be absolute,
not relative. They would be (and still are; I've verified it) relative
IFF we ALSO --load rdfs.ttl.

I'm not sure why or how this worked so far; but it's correct like this.